\name{compute_lambdas}
\alias{compute_lambdas}
\title{Compute lambda sequence}
\usage{compute_lambdas(sm)}
\arguments{
   \item{sm}{List of plde fit}
}

\description{\code{compute_lambdas} function gives the entire decreasing tuning parameter
                   sequence (\code{sm$lambda}) on the log-scale.}

\author{
JungJun Lee, Jae-Hwan Jhong, Young-Rae Cho, SungHwan Kim, Ja-yong Koo
}

\references{
JungJun Lee, Jae-Hwan Jhong, Young-Rae Cho, SungHwan Kim and Ja-Yong Koo. "Penalized Log-density Estimation Using Legendre Polynomials."
Submitted to Communications in Statistics - Simulation and Computation (2017), in revision.
}
