\name{q_lambda}
\alias{q_lambda}
\title{Compute quadratic approximation objective function}
\usage{q_lambda(sm)}
\arguments{
   \item{sm}{List of plde fit}
}

\description{\code{q_lambda} function computes
quadratic approximation of the objective function.}

\author{
JungJun Lee, Jae-Hwan Jhong, Young-Rae Cho, SungHwan Kim, Ja-yong Koo
}

\references{
JungJun Lee, Jae-Hwan Jhong, Young-Rae Cho, SungHwan Kim and Ja-Yong Koo. "Penalized Log-density Estimation Using Legendre Polynomials."
Submitted to Communications in Statistics - Simulation and Computation (2017), in revision.
}
