\name{update}
\alias{update}
\title{Update the Legendre polynomial coefficient vector}
\usage{update(sm)}
\arguments{
   \item{sm}{List of plde fit}
}

\description{\code{update} function finds the minimizer of an
             univariate quadratic approximation objective function
             for each coefficient coordinate-wise.}

\author{
JungJun Lee, Jae-Hwan Jhong, Young-Rae Cho, SungHwan Kim, Ja-yong Koo
}

\references{
JungJun Lee, Jae-Hwan Jhong, Young-Rae Cho, SungHwan Kim and Ja-Yong Koo. "Penalized Log-density Estimation Using Legendre Polynomials."
Submitted to Communications in Statistics - Simulation and Computation (2017), in revision.
}
